chore: add LICENSE file
The MIT license explicitly requires to attach the full license text to the code or any distribution. Without this the license isn't fulfilled by the project itself and thus invalid.

Also the npm package as it’s published and distributed, do not comply with this, as it doesn’t contain the license file. Effectively, without the full license text they are proprietary code and cannot be used by anyone who needs to comply with the licenses. 

Can we get this into master and publish to npm, please?